+++ RELEASE_NOTES.md
@@ -3,6 +3,9 @@
 Release notes for the AD Build System (Software Factory)
 
 ## Releases:
+* tag:1.0.17 06-Aug-2025 P. Nisperos (pnispero)
+  * Updated bs cli - removed bs generate-config and moved it over to admin configure-repo
+
 * tag:1.0.16 04-Aug-2025 P. Nisperos (pnispero)
   * Updated deployment controller for a longer timeout for a request (for large deployments)
   * Updated deployment controller paths to playbooks for different facilities